split tests into individual classes
